+++ scripts/get-release-notes.rb
@@ -9,8 +9,8 @@ end
 RELEASE_NOTES_FILE = ARGV.shift
 
 all_releases = IO.read(RELEASE_NOTES_FILE)
-next_release = /^### next release$(.+?)^### Jumi/m.match(all_releases)
+next_release = /^### Upcoming Changes$(.+?)^### Jumi/m.match(all_releases)
 unless next_release
-  raise "release notes for next release not found in: #{all_releases}"
+  raise "release notes for the upcoming release not found in: #{all_releases}"
 end
 puts next_release[1].strip