+++ scripts/prepare-release-notes.rb
@@ -12,7 +12,7 @@ RELEASE_NOTES_FILE = ARGV.shift
 RELEASE_VERSION = ARGV.shift
 
 old_release_notes = IO.read(RELEASE_NOTES_FILE)
-new_release_notes = old_release_notes.sub(/^### next release$/,
+new_release_notes = old_release_notes.sub(/^### Upcoming Changes$/,
                                           "### Jumi #{RELEASE_VERSION} (#{Date.today.strftime('%F')})")
 
 File.open(RELEASE_NOTES_FILE, 'wb') { |file|