+++ jumi-actors-generator/src/test/java/fi/jumi/actors/generator/reference/DummyListenerEventizer.java
@@ -8,14 +8,17 @@ import fi.jumi.actors.queue.MessageSender;
 
 public class DummyListenerEventizer implements Eventizer<DummyListener> {
 
+    @Override
     public Class<DummyListener> getType() {
         return DummyListener.class;
     }
 
+    @Override
     public DummyListener newFrontend(MessageSender<Event<DummyListener>> target) {
         return new DummyListenerToEvent(target);
     }
 
+    @Override
     public MessageSender<Event<DummyListener>> newBackend(DummyListener target) {
         return new EventToDummyListener(target);
     }