+++ RELEASE-NOTES.md
@@ -7,8 +7,9 @@ Release Notes
 - Jumi Actors is from this release onwards its own project with its own release cycle
 - Maven groupId was changed from `fi.jumi` to `fi.jumi.actors`
 - Upgraded to ASM 5, making the thread-safety-agent Java 8 compatible
-- New annotation processor based code generator
-  - Generates getters for event classes
+- New annotation processor based code generator. Includes the following enhancements:
+  - Getters for event classes
+  - Cleaner generated code
 
 ### Jumi Actors 0.1.196 (2012-09-19)