+++ RELEASE-NOTES.md
@@ -2,7 +2,7 @@
 Release Notes
 =============
 
-### Upcoming Changes
+### Jumi Actors 1.0.270 (2015-08-28)
 
 - Jumi Actors is from this release onwards its own project with its own release cycle
 - Maven groupId was changed from `fi.jumi` to `fi.jumi.actors`